refactor: :art: 컴포넌트 페이지별로 관리 (#57)
