+++ src/components/CommunityPage/ArticleImgCarousel.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { Navigation, Pagination } from "swiper";
+import { Pagination } from "swiper";
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/css";
 import "swiper/css/pagination";
