+++ src/components/MainPage/ViewedRecipeImgList.jsx
